Major cleanup and documentation
- redid ray-sphere intersection
- added spheres back to scene intersection
