+++ src/camera.h
@@ -1,16 +1,12 @@
 //
-// Created by s150818 on 14-2-2019.
+// Created by Tijn Bertens on 11-2-2021.
 //
 
-#ifndef RAYTRACING_SCENE_H
-#define RAYTRACING_SCENE_H
+#ifndef RAYTRACING_CAMERA_H
+#define RAYTRACING_CAMERA_H
 
-#include <random>
+#include "types.h"
 #include "math_utils.h"
-#include "color.h"
-#include "geometries.h"
-#include "ray.h"
-#include "material.h"
 
 /**
  * Representation of a camera object.
@@ -61,6 +57,9 @@ Vec3f pixelToWorldSpace(Camera camera, u32 x, u32 y, u32 width, u32 height) {
     return result;
 }
 
+/**
+ * Calculates a random world space position on a given pixel on the screen of the camera.
+ */
 Vec3f pixelToWorldSpaceRand(Camera camera, u32 x, u32 y, u32 width, u32 height) {
     Vec3f screenCenter = camera.position + (camera.viewDirection * camera.nearClippingDistance);
 
@@ -103,63 +102,4 @@ Vec3f pixelToWorldSpaceRand(Camera camera, u32 x, u32 y, u32 width, u32 height) 
     return result;
 }
 
-//  ----------------------
-//      Scene objects
-//  ----------------------
-
-/**
- * 3D sphere.
- */
-struct SphereObject {
-    Sphere sphere;              // sphere geometry
-    PBMaterial material;        // physically based material
-};
-
-/**
- * 3D triangle.
- */
-struct TriangleObject {
-    Triangle triangle;          // triangle geometry
-    PBMaterial material;        // physically based material
-};
-
-/**
- * 3D model
- */
-struct Model {
-    Mesh *mesh;                 // geometry and materials of the model
-    Matrix4x4 transform;        // model to world space transform
-};
-
-/**
- * Point light.
- */
-struct PointLight {
-    Vec3f position;          // world space position
-    Color color;                // color of emitted light
-};
-
-/**
- * Container that stores objects in a scene. Thus representing an environment filled with 3D objects.
- */
-struct Scene {
-    Color backgroundColor;
-
-    SphereObject *spheres;
-    u32 numSpheres;
-
-    TriangleObject *triangles;
-    u32 numTriangles;
-
-    Model *models;
-    u32 numModels;
-
-    PointLight *lights;
-    u32 numLights;
-};
-
-
-//  ------------------------------------------------------------------------------
-//  ------------------------------------------------------------------------------
-
-#endif //RAYTRACING_SCENE_H
+#endif //RAYTRACING_CAMERA_H
